CB-10612 Stack service updates for in-place enablement of load balanacers
Fleshes out the Stack service update_load_balancers logic. This API call is internal only, and is
intended to be called by the Environment service with an environment load balancer update is
triggered. A new flow was added to update an existing environment with load balancers. The flow
will in order:
  1. Create the load balancer database entities and save them.
  2. Create the cloud load balancer resources.
  3. Collect the load balancer cloud metadata and persist it to the database.
  4. Register the load balancers' public DNS entries with Route 53.
  5. Register the load balancers' private DNS entries with FreeIPA.
  6. Update the CM management service with the load balancer's FQDN as its frontend_url.
  7. Restart the CM server to pick up the updated setting.

Some load balancer creation logic, specifically cloud resource creation and metadata collection,
has been refactored so that it can be called independently of the wider stack creation logic.

Tested with unit tests and by running the new API commands against a stack without a load balancer
in place, and validating the database and cloud resources afterwards.